my services integrated

+++ app/api/seed/route.ts
@@ -13,21 +13,18 @@ import { dummyCategories } from '@/constants/dummyCategories';
 import { dummyUsers } from '@/constants/dummyUsers';
 import { dummyServices } from '@/constants/dummyServices';
 
+let adminID = "6618810abcccd5fae0d36f89";
+
 // main function
 export async function POST(req: Request) {
     console.log('Seeding database');
 
     // Delete everything
-    // await deleteAllCategories();
     await deleteAllServices();
     await deleteAllUsers();
 
-    // Create everything
+    // Create everything 
     await createAllUsers(dummyUsers);
-
-    // Reviews
-    // Reservations
-
 
     return NextResponse.json({ message: 'OK' });
 }
@@ -36,10 +33,8 @@ export async function POST(req: Request) {
  *  Users
  *******************************************************************/
 const createAllUsers = async (dummyUsers: any[]) => {
-    console.log('Creating all users');
-    let categoriesMade = [];
 
-    // create users
+    // create users and relationships
     for (let i = 0; i < dummyUsers.length; i++) {
         const user = dummyUsers[i];
         const createdUser = await createUser(user);
@@ -58,21 +53,25 @@ const createAllUsers = async (dummyUsers: any[]) => {
                 })
 
                 console.log('createdService:', createdService);
-
-                // create categories
-                // if (createdService) {
-                //     const category: any = dummyCategories.find(category => category._id === createdService.category);
-
-                //     // create category if it doesn't exist
-                //     if (category && categoriesMade.indexOf(category._id) === -1) {
-                //         const createdCategory: any = await createCategory({ categoryName: category.name });
-                //         categoriesMade.push(category._id);
-                //         console.log('createdCategory:', createdCategory);
-                //     }
-                // }
             }
         }
     }
+
+    // create admins' services
+    const admin = await getUserById(adminID);
+    if (admin && admin.serviceIDs.length > 0) {
+        for (let i = 0; i < admin.serviceIDs.length; i++) {
+            const service = dummyServices.find(service => service._id === admin.serviceIDs[i]);
+            const createdService = await createService({
+                service: service.params,
+                userId: adminID,
+                path: '/profile'
+            })
+
+            console.log('createdService:', createdService);
+        }
+    }
+
     return NextResponse.json({ message: 'OK', user: dummyUsers });
 }
 
@@ -81,8 +80,7 @@ const deleteAllUsers = async () => {
 
     try {
         await connectToDatabase();
-        // delete all users except for the first user
-        const deleted = await User.deleteMany({});
+        const deleted = await User.deleteMany({ _id: { $ne: adminID } });
         if (deleted) {
             console.log('Deleted all users', deleted);
         }
@@ -96,16 +94,6 @@ const deleteAllUsers = async () => {
 /*******************************************************************
  *  Services
  *******************************************************************/
-const createAllServices = async (dummyServices: any[]) => {
-    console.log('Creating all services');
-    // create services
-    for (let i = 0; i < dummyServices.length; i++) {
-        const service = dummyServices[i];
-        await createService(service);
-    }
-    return NextResponse.json({ message: 'OK', service: dummyServices });
-}
-
 const deleteAllServices = async () => {
     console.log('Deleting all services');
     try {
@@ -117,35 +105,6 @@ const deleteAllServices = async () => {
         return NextResponse.json({ message: 'OK', service: deleted })
     } catch (error) {
         console.error('Error deleting all services:', error);
-        return NextResponse.json({ message: 'Error', error: error })
-    }
-}
-
-/*******************************************************************
- *  Categories
- *******************************************************************/
-const createAllCategories = async (dummyCategories: any[]) => {
-    console.log('Creating all categories');
-    // create categories
-    for (let i = 0; i < dummyCategories.length; i++) {
-
-        const categoryName = dummyCategories[i].name;
-        await createCategory({ categoryName });
-    }
-    return NextResponse.json({ message: 'OK', category: dummyCategories });
-}
-
-const deleteAllCategories = async () => {
-    console.log('Deleting all categories');
-    try {
-        await connectToDatabase();
-        const deleted = await Category.deleteMany({});
-        if (deleted) {
-            console.log('Deleted all categories', deleted);
-        }
-        return NextResponse.json({ message: 'OK', category: deleted })
-    } catch (error) {
-        console.error('Error deleting all categories:', error);
         return NextResponse.json({ message: 'Error', error: error })
     }
 }
